requirements.txt
Pillow
Django
Markdown
django-mdeditor
PyMySQL